requirements.txt
pdf2image
pytesseract
psycopg2-binary
numpy
openai
pandas
msal
requests
tqdm
psycopg2-binary
python-docx
python-dotenv
flask
flask_session
openpyxl
striprtf
tiktoken
waitress
